ml_service/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
google-generativeai>=0.3.2
google-cloud-aiplatform>=1.65.0
python-dotenv>=1.0.1
pydantic>=2.6.1
firebase-admin>=6.5.0
scikit-learn>=1.5.1
numpy>=1.26.4